refactor: split agent_reflect into CoT and Direct agent modules
* Move CoTAgent implementation into agent_cot.py and DirectAgent into agent_direct.py
* Update relevant imports and references across the codebase
* Add descriptive comment in renamed agent_reflect.py file
